Fixed bug causing AI to damage Player when player is not physically hit
AI was still hitting the player due to the player being within the enemy's trigger. Removing and reworking the player detection to the fov raycasting fixed this issue, but now the player is unable to detect the enemies for lock-on. Will have to rework the player to have a similar ray-cast detection;